entities

+++ src/main/java/com/max/TaskFlow/model/Task.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class Task {
     @Id()
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int taskId;
+    private Long taskId;
     private String name;
     private String description;
     private LocalDateTime dateCreated;
